[#8] Run Jetpack Compose instrument tests with Robolectric https://medium.com/@sebaslogen/blazing-fast-compose-tests-with-robolectric-b059f5471495
