+++ android/src/test/resources/robolectric.properties
@@ -1,0 +1,2 @@
+# https://github.com/robolectric/robolectric/issues/6593
+instrumentedPackages=androidx.loader.content
